refactor: separate events-board component
