+++ src/components/shared/events-modal/events-board/index.js
@@ -1,0 +1,3 @@
+import EventsBoard from './events-board';
+
+export default EventsBoard;
